Additional code and scripts copied from plaso project
