+++ l2tpreg/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+"""Log2timeline preg module."""
+
+__version__ = '20170512'
