+++ l2tpreg/definitions.py
@@ -1,0 +1,18 @@
+# -*- coding: utf-8 -*-
+"""The definitions."""
+
+REGISTRY_FILE_TYPE_NTUSER = u'NTUSER'
+REGISTRY_FILE_TYPE_SAM = u'SAM'
+REGISTRY_FILE_TYPE_SECURITY = u'SECURITY'
+REGISTRY_FILE_TYPE_SOFTWARE = u'SOFTWARE'
+REGISTRY_FILE_TYPE_SYSTEM = u'SYSTEM'
+REGISTRY_FILE_TYPE_UNKNOWN = u'UNKNOWN'
+REGISTRY_FILE_TYPE_USRCLASS = u'USRCLASS'
+
+REGISTRY_FILE_TYPES = frozenset([
+    REGISTRY_FILE_TYPE_NTUSER,
+    REGISTRY_FILE_TYPE_SAM,
+    REGISTRY_FILE_TYPE_SECURITY,
+    REGISTRY_FILE_TYPE_SOFTWARE,
+    REGISTRY_FILE_TYPE_SYSTEM,
+    REGISTRY_FILE_TYPE_USRCLASS])
